Create single-use trait to collect HydeKernel string helpers
