+++ packages/framework/src/Foundation/Concerns/ImplementsStringHelpers.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Hyde\Framework\Foundation\Concerns;
+
+use Illuminate\Support\Str;
+
+/**
+ * @internal Single-use trait for the HydeKernel class.
+ * @see \Hyde\Framework\HydeKernel
+ */
+trait ImplementsStringHelpers
+{
+    public function makeTitle(string $slug): string
+    {
+        $alwaysLowercase = ['a', 'an', 'the', 'in', 'on', 'by', 'with', 'of', 'and', 'or', 'but'];
+
+        return ucfirst(str_ireplace(
+            $alwaysLowercase,
+            $alwaysLowercase,
+            Str::headline($slug)
+        ));
+    }
+}
